options and basic procedures performance refactored
