+++ src/basic_procedures.h
@@ -1,0 +1,201 @@
+#pragma once
+
+
+#include <string>
+#include <utility>
+
+#include "oscillation/oscillation_basic.h"
+#include "fft/fftw_impl.h"
+#include "periods/periods_base.h"
+#include "oscillation/cut_oscillation_file.h"
+#include "oscillation/wt_oscillation.h"
+#include "flow/parse_ptl.h"
+#include "analize_coefficients/dynamic_coefficients.h"
+
+
+namespace basic_procedures
+{
+    enum ErrorCodes
+    {
+        SUCCESS     = 0,
+        FAIL        = 1,
+        UNEXPECTED  = 2
+    };
+
+
+    /**
+    *    Data loaded as Oscillation object. So the the initial data is y and t. And while Oscillation constructed
+    *    y' and y'' is calculating.
+    *
+    *    Basically procedure finds periods of the signal and saves it in specific manner.
+    *    Each period splits into forward a backward movement
+    *    (todo) after that gnuplot script creating.
+    */
+    ErrorCodes performProcedurePeriods(const std::string& fileName,
+                                       const std::string& fileName2,
+                                       const std::vector<double>& extraArguments)
+    {
+        std::cout << "periods procedure performing...\n";
+
+        std::ifstream fin();
+
+        Oscillation D;
+        D.loadFile(fileName);
+        if(0 == D.size())
+        {
+           std::cerr << "Empty Oscillation data! Aborting procedure\n";
+           return FAIL;
+        }
+
+
+        std::cout << "extra arguments size = " << extraArguments.size() << std::endl;
+
+        int mergeHalfPeriods = 1;
+
+        if (extraArguments.size() > 1)
+        {
+            mergeHalfPeriods = extraArguments.at(1);
+            std::cout << "mergeHalfPeriods = " << mergeHalfPeriods <<  "\n";
+        }
+
+
+        std::vector<Oscillation> periodsList;
+
+        periodsList = periods::splitPeriods(D);
+
+        std::cout << "Data split into " << periodsList.size() << "periods\n";
+
+        std::string outputFileName = std::string();
+
+        std::ofstream fout;
+
+        std::ofstream fout2(fileName + ".periods");
+
+        for(size_t i = 0; i < periodsList.size();)
+        {
+            outputFileName = fileName2 + "_" + std::to_string(i);
+
+            if (mergeHalfPeriods > 1)
+            {
+                outputFileName += "_" + std::to_string(i + mergeHalfPeriods);
+            }
+
+            outputFileName += "_pi";
+
+            fout.open(outputFileName);
+
+            for (int halfperiodsToFile = 0; (halfperiodsToFile < mergeHalfPeriods && i < periodsList.size()); halfperiodsToFile++)
+            {
+                fout << periodsList.at(i);
+
+                if (i <  periodsList.size()-2)
+                {
+                    fout2 << periodsList.at(i).time.at(0) << "\t"
+                            << periodsList.at(i).angle.at(0) << "\t"
+                            << periodsList.at(i).dangle.at(0) << "\t"
+                            << periodsList.at(i).ddangle.at(0) << "\t"
+                            << periodsList.at(i+2).time.at(0) - periodsList.at(i).time.at(0) << "\t"
+                            << 1/(periodsList.at(i+2).time.at(0) - periodsList.at(i).time.at(0)) << "\n";
+                }
+
+                i++;
+            }
+
+
+            fout.close();
+        }
+        fout2.close();
+
+        return SUCCESS;
+    }
+
+    ErrorCodes performProcedureFlow(const std::string& fileName)
+    {
+        wt_flow::Flow flow;
+
+        if( wt_flow::parsePTLfile(fileName + ".ptl", flow, 5))
+        {
+            flow.saveFile(fileName + ".flow");
+            std::cout << "File " << fileName << " parsed: success\n";
+            return SUCCESS;
+        }
+
+        std::cerr << "File " << fileName << " parsed: failed\n";
+
+        return FAIL;
+    }
+
+    ErrorCodes performProcedureCutFile( const std::string& initialFile,
+                                        const double timeFrom,
+                                        const double timeTo,
+                                        const std::string& finalFile)
+    {
+        std::cout << "cut file performing\n";
+
+        if (!Oscillation_files::cut_file(initialFile, timeFrom , timeTo, finalFile))
+        {
+            std::cerr << "cut file failed!/n";
+            return FAIL;
+        }
+
+        return SUCCESS;
+    }
+
+
+    ErrorCodes performProceduereProcessingWtData(const std::string& fileName)
+    {
+        std::cout << "wt test processing mode performing\n";
+
+        /// FLOW->
+        wt_flow::Flow flow;
+        if (!flow.loadFile((fileName + "_flow")))
+            return FAIL;
+
+        flow.calculateFlow();
+        flow.print();
+        /// FLOW<-
+
+        /// MODEL->
+        Model model;
+        if (!model.loadFile((fileName + "_model")))
+            return FAIL;
+
+        model.print();
+        /// MODEL<-
+
+        /// OSCILLATION->
+        Oscillation oscillation;
+        if (!oscillation.loadFile(fileName + "_osc"))
+            return FAIL;
+        /// OSCILLATION<-
+
+
+        WtOscillation wtTest(oscillation, flow, model);
+
+        wtTest.getMz();
+        wtTest.saveMzData(fileName + "_mz");
+        //todo check
+
+
+        wtTest.getMzAmplitudeIndexes();
+        wtTest.saveMzAmplitudeData(fileName + "_mz_amplitude");
+
+
+        return SUCCESS;
+    };
+
+
+    void testFunc()
+    {
+        Oscillation A("A");
+        std::vector<Oscillation> arr_osc;
+        arr_osc.push_back(A);
+        arr_osc[0].write("A_osc");
+    }
+
+    /*
+    TODO
+     фильтрация, опции, нормальный функционал, деление на логику и работу с файлами
+    */
+
+}
