+++ src/options.cpp
@@ -1,17 +1,19 @@
 #include "options.h"
 
-
-bool Options::parse_options(int ac, char* av[])
+namespace options
 {
-    po::store(po::command_line_parser(ac, av).options(*descr).run(), m_vm);
-    po::notify(m_vm);
-
-    if(exist("help"))
+    bool Options::parse_options(int ac, char* av[])
     {
-        //todo
-        //show_help();
+        po::store(po::command_line_parser(ac, av).options(*descr).run(), m_vm);
+        po::notify(m_vm);
+
+        if(exist("help"))
+        {
+            //todo
+            //show_help();
+            return true;
+        }
+
         return true;
     }
-
-    return true;
 }
